+++ Libraries/XulbuX/CHANGELOG.md
@@ -17,6 +17,8 @@
 
 ## 31.10.2024 `v1.5.4`
 * Made the `blend()` method from all the color types modify the *`self`* object as well as returning the result
+* Added a new function <code>normalize_spaces(*code*) -> *str*</code> to `Code`
+* Added new doc-strings to `xx_code`
 
 ## 30.10.2024 `v1.5.3`
 * Restructured the values in `_consts_.py`